chore: rename 'validate' to 'header_validation' for clarity.
