+++ amaru/src/consensus/header_validation.rs
