+++ amaru/src/consensus/mod.rs
@@ -1,4 +1,4 @@
-pub mod validate;
+pub mod header_validation;
 
 pub type RawBlock = Vec<u8>;
 pub type Point = pallas_network::miniprotocols::Point;
